third_party/blink/web_tests/platform/generic/fast/css/font-face-download-error-expected.txt

If no font resources are available for an @font-face rule due to download errors, the rule must be ignored but others must remain in effect.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

iiiiiiiiii test_0_0
iiiiiiiiii ref_0_0
jjjjjjjjjj test_0_1
jjjjjjjjjj ref_0_1

iiiiiiiiii test_1_0
iiiiiiiiii ref_1_0
jjjjjjjjjj test_1_1
jjjjjjjjjj ref_1_1

iiiiiiiiii test_2_0
iiiiiiiiii ref_2_0
jjjjjjjjjj test_2_1
jjjjjjjjjj ref_2_1

iiiiiiiiii test_3_0
iiiiiiiiii ref_3_0
jjjjjjjjjj test_3_1
jjjjjjjjjj ref_3_1


PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById('test_0_0').offsetWidth is document.getElementById('ref_0_0').offsetWidth
PASS document.getElementById('test_0_1').offsetWidth is document.getElementById('ref_0_1').offsetWidth
PASS document.getElementById('test_1_0').offsetWidth is document.getElementById('ref_1_0').offsetWidth
PASS document.getElementById('test_1_1').offsetWidth is document.getElementById('ref_1_1').offsetWidth
PASS document.getElementById('test_2_0').offsetWidth is document.getElementById('ref_2_0').offsetWidth
PASS document.getElementById('test_2_1').offsetWidth is document.getElementById('ref_2_1').offsetWidth
PASS document.getElementById('test_3_0').offsetWidth is document.getElementById('ref_3_0').offsetWidth
PASS document.getElementById('test_3_1').offsetWidth is document.getElementById('ref_3_1').offsetWidth
